chore: break observation interpretation file
